Add save/load to 6 independent egs mode
It will now detect slide and pot changes (for enabled channels). Once a
change has occurred, and at least 5 seconds have passed since the last
change, the state will be written to flash memory so that the envelopes
can survive a power cycle.

There is more work to be done around the mode's static variables - they
initialize only once upon startup (to dummy values). This both makes it
detect initial slider positions on startup as moves (if not at zero), as
well as makes it not reset the initial slider positions when changing
modes.