Add Envelope Mode
This moves all of the envelope mode logic into its own class that wraps
the envelope manager. This allows the UI to get a pointer to it and
invoke reinitializing logic when changing modes.

The envelopes now start in the correct state both upon bootup and upon
entering the independent envelopes mode. That is, it starts with the
first channel active, all envelopes loaded from save state, and all
sliders disabled, regardless of position. Sliders will only enable once
the user has moved them.