Refactor envelope modes
This renames EnvelopeManager to EnvelopeMode and refactors it such that it contains all of the envelope mode processing logic. This allows mode changes (detected during ui polling) to re-initialize the variables used in these modes.

This makes switching to these modes more robust and consistent as they no longer rely on static variables that are set to arbitrary values upon boot, and left at arbitrary values upon switching modes.