Add save/load to 6 independent egs mode (#65)
* Add save/load to 6 independent egs mode

It will now detect slide and pot changes (for enabled channels). Once a
change has occurred, and at least 5 seconds have passed since the last
change, the state will be written to flash memory so that the envelopes
can survive a power cycle.

There is more work to be done around the mode's static variables - they
initialize only once upon startup (to dummy values). This both makes it
detect initial slider positions on startup as moves (if not at zero), as
well as makes it not reset the initial slider positions when changing
modes.

* Remove debug leds and dead code

Saving and loading are _generally_ working.

There is still more to do when switching modes and startup.

There should be a more explicit reinit for the mode that sets all the state and does not rely upon static variables.

* Rename manager mode

Also make naming more consistent with followup PR

* Cleanup

* Push comment upstream

* Rename envelope_mode back to envelope_manager

* More EnvelopeMode->EnvelopeManager renames

* Address feedback

Fix ranges for slider and pot value storing (should get double the
precision now) and convert IEGParams to an enum.