Use slider leds to indicate channel status
For the non-active channels, leds will be either on or off to indicate
whether the slider is enabled (or not).

For the active channel, the led will pulse. It will pulse more
frequently and for a longer duration if slider is enabled than if it is
disabled.

This replaces the previous logic that showed which channel was active by
marking the main led yellow (and only when the envelope wasn't active).
This was both confusing (as yellow also indicated a specific envelope
stage) and might never be shown if the envelope is retriggered prior to
completing.

Ideally we would be performing a smoother fade, but the leds on the
sliders seem to only support binary on/off.